add packages and classes for specific tasks
