+++ src/main/java/src/main/ssagit/Task.java
@@ -1,0 +1,2 @@
+package src.main.ssagit;public class Task {
+}
