+++ src/main/java/src/main/ssagit/TaskType.java
@@ -1,0 +1,2 @@
+package src.main.ssagit;public enum TaskType {
+}
