+++ src/main/java/ssagit/TaskType.java
@@ -1,0 +1,7 @@
+package ssagit;
+
+public enum TaskType {
+    TODO,
+    DEADLINE,
+    EVENT
+}
